+++ src/blogmore/linter.py
@@ -37,6 +37,7 @@ class Linter:
         self.parser = PostParser(
             site_url=site_config.site_url,
             default_show_toc=site_config.show_toc,
+            default_show_toc_inline=site_config.show_toc_inline,
         )
         self.errors = 0
         self.warnings = 0